Restructured App component and fixed path of projects page
