+++ frontend/src/App.js
@@ -1,24 +1,29 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import './App.css';
+
+// --------------- LAYOUT
 import Navbar from './components/Layout/Navbar';
-import Contributors from './components/Contributors';
 import Footer from './components/Layout/Footer';
+import ScrollToTopButton from './components/scrolltotopButton';
+
+// --------------- PAGES
+import Contributors from './components/Contributors';
 import EventCreation from './components/common/EventCreation';
 import AboutPage from './components/AboutPage';
 import EventsPage from './Pages/Events/EventsPage';
 import HackathonPage from './Pages/Hackathons/HackathonPage';
-import ProjectGallery from './components/ProjectGallery';
-import ScrollToTopButton from './components/scrolltotopButton';
+import ProjectsPage from './Pages/Projects/ProjectsPage';
 import NotFound from './components/NotFound';
-import './App.css';
 
-// Import Auth components
+
+// --------------- AUTH PAGES
 import Login from './components/auth/Login';
 import Signup from './components/auth/Signup';
 import Unauthorized from './components/auth/Unauthorized';
 import ProtectedRoute from './components/auth/ProtectedRoute';
 import { AuthProvider } from './context/AuthContext';
 
-// Import Dashboard components
+// --------------- DASHBOARD PAGES
 import Dashboard from './components/Dashboard';
 import AdminDashboard from './components/admin/AdminDashboard';
 import HomePage from './Pages/Home/HomePage';
@@ -37,7 +42,7 @@ function App() {
               <Route path="/" element={<HomePage />} />
               <Route path="/events" element={<EventsPage />} />
               <Route path="/hackathons" element={<HackathonPage />} />
-              <Route path="/projects" element={<ProjectGallery />} />
+              <Route path="/projects" element={<ProjectsPage />} />
               <Route path="/contributors" element={<Contributors />} />
 
               {/* Protected route for event creation */}
@@ -83,7 +88,7 @@ function App() {
           <Footer />
         </div>
       </Router>
-    </AuthProvider>
+    </AuthProvider >
   );
 }
